+++ drel/requests/__init__.py
@@ -1,1 +1,1 @@
-from .api import log, post
+from .api import log, post, get